+++ examples/world.rs
@@ -16,7 +16,10 @@ fn main() {
         // bevy_ecs_tilemap and bevy_ecs_tiled main plugins
         .add_plugins(TilemapPlugin)
         .add_plugins(TiledMapPlugin::default())
+        // Examples helper plugins: for this example, contains the logic to move the camera
         .add_plugins(helper::HelperPlugin)
+        // Enable debug informations
+        .add_plugins(TiledDebugPluginGroup)
         // Add our systems and run the app!
         .add_systems(Startup, startup)
         .add_systems(Update, (input, text_update_system))